Moved app engine datastore setup to abstractTestCase
