+++ src/test/edu/iastate/music/marching/attendance/test/integration/AbstractTestCase.java
@@ -1,12 +1,29 @@
 package edu.iastate.music.marching.attendance.test.integration;
 
+import org.junit.After;
+import org.junit.Before;
+
+import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
+import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
 import com.google.code.twig.ObjectDatastore;
 import com.google.code.twig.annotation.AnnotationObjectDatastore;
 
 import edu.iastate.music.marching.attendance.controllers.DataTrain;
 
 public class AbstractTestCase {
-	
+
+	private final LocalServiceTestHelper helper = new LocalServiceTestHelper(
+			new LocalDatastoreServiceTestConfig());
+
+	@Before
+	public void setUp() {
+		helper.setUp();
+	}
+
+	@After
+	public void tearDown() {
+		helper.tearDown();
+	}
 
 	protected DataTrain getDataTrain() {
 		return DataTrain.getAndStartTrain();
@@ -15,5 +32,5 @@ public class AbstractTestCase {
 	protected ObjectDatastore getObjectDataStore() {
 		return new AnnotationObjectDatastore(false);
 	}
-	
+
 }
