+++ src/test/edu/iastate/music/marching/attendance/test/integration/MobileDataUpload.java
@@ -1,17 +1,13 @@
 package edu.iastate.music.marching.attendance.test.integration;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
 import java.util.Calendar;
 import java.util.List;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
-import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
 import com.google.code.twig.ObjectDatastore;
 import com.google.code.twig.annotation.AnnotationObjectDatastore;
 
@@ -33,19 +29,6 @@ public class MobileDataUpload extends AbstractTestCase {
 			+ "tardyStudent&split&Daniel&split&Stiner&split&stiner&split&2012-05-03&split&0107&split&|&split&null&newline&"
 			+ "tardyStudent&split&Daniel&split&Stiner&split&stiner&split&2012-05-03&split&0108&split&|&split&null&newline&"
 			+ "tardyStudent&split&Yifei&split&Zhu&split&zyf&split&2012-05-03&split&0108&split&|&split&4&newline&";
-
-	private final LocalServiceTestHelper helper = new LocalServiceTestHelper(
-			new LocalDatastoreServiceTestConfig());
-
-	@Before
-	public void setUp() {
-		helper.setUp();
-	}
-
-	@After
-	public void tearDown() {
-		helper.tearDown();
-	}
 
 	@Test
 	public void simpleAbsenceInsertionThroughController() {
